Added manifest.json: Contains add-on metadata, as defined by the WebExtensions specification (https://developer.mozilla.org/en-US/Add-ons/WebExtensions/manifest.json). Generated from manifest.yaml
Added manifest.yaml: JSON files are a pain to work with manually, especially if when long string values are required. YAML is much better for this, but WebExtensions won't except manifest file in YAML. The compromise is to generate manifest.json from manifest.yaml. See comments in manifest.yaml for guidance.